requirements.txt
hydra-core==1.3.2
lightning==2.4.0
torch==2.4.0
torchvision==0.19.0
